move rname to email processing to a bit
