+++ octopoes/bits/email_from_soa/bit.py
@@ -1,0 +1,10 @@
+from bits.definitions import BitDefinition
+from octopoes.models.ooi.dns.records import DNSSOARecord
+
+BIT = BitDefinition(
+    id="email-from-soa",
+    consumes=DNSSOARecord,
+    parameters=None,
+    module="bits.email_from_soa.email_from_soa",
+    min_scan_level=0,
+)
